pyproject.toml
[tool.poetry]
name = "stv-python-client"
version = "0.1.0"
description = ""
authors = ["Moritz Makowski <moritz.makowski@tum.de>"]

[tool.poetry.dependencies]
python = "3.10"
mysql-connector-python = "^8.0.29"

[tool.poetry.dev-dependencies]
black = "22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
